fix: stop favicon requests for private chat-link hosts
Chat markdown and mobile feed links asked Google for a favicon of every
host, including loopback, RFC1918, Tailscale, and .local names. Shared
classification now refuses those hosts so the globe fallback is used
instead.

Adapted from pingdotgg/t3code#5838.
